Rename Code_Checking/Code_Checking_Program.cpp to Code_Checking.cpp
As you write your code, it is very easy to leave out a parenthesis, bracket, or curly brace. To help solve this problem, this simple program will check to make sure all the parentheses, brackets, and curly braces lineup.

The program prompts the user for a file name of a program. Once the file is added, the program will look to make sure everything matches up and if there is an imbalance, it reports the line number, the symbol that was found, and the symbol that was expected.